@@ -5,4 +5,4 @@ require (
 	golang.org/x/sys v0.0.0-20200728102440-3e129f6d46b1 // indirect
 )
 
-go 1.14
+go 1.16
